refactor: reduce complexity in joke-page
